==> 1. Create Price-Return_data.ipynb <==
# 1. Import library

import yfinance as yf
from datetime import date, datetime
import pandas as pd

# 2. Download NAV data

Download NAV data from here:
https://www.franklintempleton.co.uk/our-funds/price-and-performance/products/91383/AUD/ftgf-clearbridge-us-equity-sustainability-leaders-fund/IE00BZ1G4Q59#view-historical-pricing

# 3. Import NAV data

data_1 = pd.read_excel('fund_price.xlsx', index_col='Date')
data_1.columns = ['Fund']
data_1.head()

data_1.info()

# 4. Manage datetime index

data_1.index = pd.to_datetime(data_1.index, format = '%d/%m/%Y')
data_1.sort_values(by='Date', ascending = True, inplace = True)
data_1['Fund'] = [x.strip('$') for x in data_1['Fund']]
data_1['Fund'] = data_1['Fund'].astype(float)

data_1.head()

#data_1.tail()

data_1.info()

# 5. Download All holding in FTGF ClearBridge US Equity Sustainability Leaders Fund

Download All Holding Stock from here : https://www.franklintempleton.co.uk/our-funds/price-and-performance/products/91383/AUD/ftgf-clearbridge-us-equity-sustainability-leaders-fund/IE00BZ1G4Q59#view-historical-pricing

# 6. Import data

data_2 = pd.read_excel('stock_list.xlsx')

data_2.head()

# 7. Select only "Ticker" columns from the data frame

stock_list = data_2['Ticker'].to_list()

#stock_list

# stock_list = ['AAPL',  'APTV',  'BAC', 'BALL',  'BEP', 'BKNG', 'BMRN', 'COST', 'CVS', 'DE', 'DHR', 'DIS', 
#  'ECL',  'EQIX', 'ETN', 'GOOG', 'HD', 'HIG', 'MKC',  'MS', 'MSFT', 'NEP',  'PGR', 'PLD',  'TEL', 'TREX', 'TT', 'UNH', 'V', 'WSM']

# 8. Select Ticker from Top 3 or 4 in each sector

stock_list = ['AAPL',  'APTV', 'BAC', 'BALL',  'BEP', 'BKNG', 'BMRN', 'COST', 'CVS', 'DE', 'ETN', 'GOOGL', 'HD', 'HIG',  'MS', 'MSFT', 'PLD',  'TT', 'UNH', 'V']

# 9. Download adjusted close of each tiker (same period as the Fund)

#data_1.index

data_3 = yf.download(stock_list, start = '2015-09-30', end = '2023-03-04')
data_3 = data_3['Adj Close']
data_3.index = pd.to_datetime(data_3.index).strftime('%Y-%m-%d')
data_3.index = pd.to_datetime(data_3.index)

data_3.head()

data_3.tail()

data_3.info()

# 10. Combine stock price and fund price

price = pd.concat([data_1, data_3.reindex(data_1.index)], axis=1)
price = price.dropna()

price.head()

price.info()

price.tail()

# 11. Daily price to Monthly price

price_m = price.asfreq('M').ffill()

price_m

# 12. Monthly return

return_m = price_m.pct_change()
for i in ret.columns:
    return_m.rename(columns={i: "ret_"+i}, inplace=True)

return_m

return_m

# 13. Save monthly price and return to CSV

price_m.to_csv('price_data.csv')

return_m.to_csv('return_data.csv')
