==> 2. Create Climate data.ipynb <==
# 1. Import library

from datetime import date, datetime
import pandas as pd

Download data from : https://www.iea.org/data-and-statistics/data-product/weather-for-energy-tracker

You can study detail in each climate data as link above

# 2. Import data for check data structure

cdd_18  = pd.read_csv('test_CDD_18.csv')

cdd_18.head(3)

cdd_18.info()

# 3. Create datetime index for the data

dt_index = pd.date_range('01-01-10', periods=156, freq='M')

dt_index

# 4. Download all climate data

### 4.1 Download all climate data

cdd_18  = pd.read_csv('CDD_18.csv', index_col='Country')
cdd_21  = pd.read_csv('CDD_21.csv', index_col='Country')
irradiance  = pd.read_csv('Global_horizontal_irradiance.csv', index_col='Country')
hdd_16  = pd.read_csv('HDD_16.csv', index_col='Country')
hdd_18  = pd.read_csv('HDD_18.csv', index_col='Country')
heat  = pd.read_csv('Heat_index.csv', index_col='Country')
humidity  = pd.read_csv('Relative_humidity.csv', index_col='Country')
temperature = pd.read_csv('Temperature.csv', index_col='Country')
precipitation  = pd.read_csv('Total_precipitation.csv', index_col='Country')

### 4.2 Change column name for the top 8 industrial countries

for i in cdd_18.columns:
    cdd_18.rename(columns = {i:'cdd_18-'+i}, inplace = True)

cdd_18.head(1)

for i in cdd_21.columns:
    cdd_21.rename(columns = {i:'cdd_21-'+i}, inplace = True)

cdd_21.head(1)

for i in irradiance.columns:
    irradiance.rename(columns = {i:'irradiance-'+i}, inplace = True)

irradiance.head(1)

for i in hdd_16.columns:
    hdd_16.rename(columns = {i:'hdd_16-'+i}, inplace = True)

hdd_16.head(1)

for i in hdd_18.columns:
    hdd_18.rename(columns = {i:'hdd_18-'+i}, inplace = True)

hdd_18.head(1)

for i in heat.columns:
    heat.rename(columns = {i:'heat-'+i}, inplace = True)

heat.head(1)

for i in humidity.columns:
    humidity.rename(columns = {i:'humidity-'+i}, inplace = True)

humidity.head(1)

for i in temperature.columns:
    temperature.rename(columns = {i:'temperature-'+i}, inplace = True)

temperature.head(1)

for i in precipitation.columns:
    precipitation.rename(columns = {i:'precipitation-'+i}, inplace = True)

precipitation.head(1)

# 5. Combine all climate data in one dataframe

climate = pd.concat([cdd_18,
                     cdd_21.reindex(cdd_18.index),
                    irradiance.reindex(cdd_18.index),
                    hdd_16.reindex(cdd_18.index),
                    hdd_18.reindex(cdd_18.index),
                    heat.reindex(cdd_18.index),
                    humidity.reindex(cdd_18.index),
                    temperature.reindex(cdd_18.index),
                    precipitation.reindex(cdd_18.index)], axis=1)

climate

# 6. Datetime Index

climate.index = dt_index

climate

# 7. Slice data for required period

climate = climate.loc['2015-09':]

climate

# 8. Change index name

climate.index.name = 'Date'

# 9. Save to CSV

climate.to_csv('climate_data.csv')
