==> 3. FamaFrench Data.ipynb <==
import pandas as pd
import getFamaFrenchFactors as gff

ff3_monthly = gff.famaFrench3Factor(frequency='m')
ff3_monthly.rename(columns={"date_ff_factors": 'Date'}, inplace=True)
ff3_monthly.set_index('Date', inplace=True)

ff3_monthly

ff3_monthly = ff3_monthly['2015-09':'2022-12'] # start refers to start of Fund price data # end refers to climate data
ff3_monthly

ff3_monthly.to_csv('fama_data.csv')
