==> 3. FamaFrench5 Data.ipynb <==
# 1. Import Library

import pandas as pd
import getFamaFrenchFactors as gff

# 2. Download FamaFrench 5 Factors data

ff5_monthly = gff.famaFrench5Factor(frequency='m')
ff5_monthly.rename(columns={"date_ff_factors": 'Date'}, inplace=True)
ff5_monthly.set_index('Date', inplace=True)

ff5_monthly

# 3. Slice data for required period

ff5_monthly = ff5_monthly['2015-09':'2022-12'] # start refers to start of Fund price data # end refers to climate data
ff5_monthly

# 4. Save to CSV

ff5_monthly.to_csv('fama_data.csv')
