==> 4. Create All data.ipynb <==
import pandas as pd

ret = pd.read_csv('return_data.csv', index_col='Date', parse_dates=True)
ret

ret.info()

climate = pd.read_csv('climate_data.csv', index_col='Date', parse_dates=True)
climate

fmf = pd.read_csv('fama_data.csv', index_col='Date', parse_dates=True)
fmf

all_data = pd.concat([fmf, 
                      climate, 
                      ret.reindex(fmf.index)], axis=1)

all_data

all_data.info()

all_data.to_csv('all_data.csv')
