==> 5. ESG Score.ipynb <==
import pandas as pd
import yesg # https://pypi.org/project/yesg/

# stock_list = ['AAPL',  'APTV',  'BAC', 'BALL',  'BEP', 'BKNG', 'BMRN', 'COST', 'CVS', 'DE', 'DHR', 'DIS',
#               'ECL',  'EQIX', 'ETN', 
#               'GOOG', 
#               'HD', 'HIG', 'MKC',  'MS', 'MSFT', 
#               #'NEP',  
#               'PGR', 'PLD',  'TEL', 
#               #'TREX', 
#               'TT', 'UNH', 'V', 
#               #'WSM']

# stock_list = ['AAPL',  'APTV',  'BAC', 'BALL',  'BEP', 'BKNG', 'BMRN', 'COST', 'CVS', 'DE', 'DHR', 'DIS', 
#  'ECL',  'EQIX', 'ETN', 'GOOGL', 'HD', 'HIG', 'MKC',  'MS', 'MSFT',  'PGR', 'PLD',  'TEL', 'TT', 'UNH', 'V']

#stock_list = ['AAPL',  'APTV', 'BAC', 'BALL',  'BEP', 'BKNG', 'BMRN', 'COST', 'CVS', 'DE', 'ETN', 'GOOGL', 'HD', 'HIG',  'MS', 'MSFT', 'PLD',  'TT', 'UNH', 'V']

stock_list = ['AAPL','CVS','ETN','HD','HIG','MS','PLD','UNH']

df = pd.DataFrame()
for i in stock_list:
    df[i+'_Total-Score'] = yesg.get_historic_esg(i)['Total-Score']
    df[i+'_E-Score'] = yesg.get_historic_esg(i)['E-Score']
    df[i+'_S-Score'] = yesg.get_historic_esg(i)['S-Score']
    df[i+'_G-Score'] = yesg.get_historic_esg(i)['G-Score']

df

df.info()

columns = ['BALL_Total-Score', 'BALL_E-Score', 'BALL_S-Score', 'BALL_G-Score', 
           'BKNG_Total-Score', 'BKNG_E-Score', 'BKNG_S-Score', 'BKNG_G-Score', 
           'TT_Total-Score', 'TT_E-Score', 'TT_S-Score', 'TT_G-Score',]
df.drop(columns, inplace=True, axis=1)

df

df = df.loc['2015-08-31':]

df

df.mean()
