==> 5. FamaFrench_and_Climate_factor_with_XGBoost.ipynb <==
# 1. Import library

import pandas as pd
import numpy as np
from xgboost import XGBRegressor
from sklearn.model_selection import (GridSearchCV, TimeSeriesSplit, train_test_split)
from sklearn.metrics import (mean_squared_error, mean_absolute_error)

# 2. Import file

all_data  = pd.read_csv('all_data.csv', index_col='Date', parse_dates=True)

all_data

# 3. Data cleaning

all_data.fillna( method ='ffill', inplace = True)

all_data.isna().sum().sum() # nan from calculate return at first row

all_data.dropna(inplace=True)

all_data.isna().sum().sum()

# 4. Determine features

features = ['Mkt-RF', 'SMB', 'HML', 'RMW', 'CMA', 
                   
                   'cdd_18-France','cdd_18-Germany', 'cdd_18-India', 'cdd_18-Indonesia', 'cdd_18-Italy', 'cdd_18-Japan', 'cdd_18-Korea', 'cdd_18-China', 
                   
                   'cdd_21-France', 'cdd_21-Germany', 'cdd_21-India', 'cdd_21-Indonesia', 'cdd_21-Italy', 'cdd_21-Japan', 'cdd_21-Korea', 'cdd_21-China', 
                   
                   'irradiance-France', 'irradiance-Germany', 'irradiance-India', 'irradiance-Indonesia', 'irradiance-Italy', 'irradiance-Japan', 'irradiance-Korea',
                   'irradiance-China', 
                   
                   'hdd_16-France', 'hdd_16-Germany', 'hdd_16-India', 'hdd_16-Indonesia', 'hdd_16-Italy', 'hdd_16-Japan', 'hdd_16-Korea', 'hdd_16-China', 
                   
                   'hdd_18-France', 'hdd_18-Germany', 'hdd_18-India', 'hdd_18-Indonesia', 'hdd_18-Italy', 'hdd_18-Japan', 'hdd_18-Korea', 'hdd_18-China', 
                   
                   'heat-France', 'heat-Germany', 'heat-India', 'heat-Indonesia', 'heat-Italy', 'heat-Japan', 'heat-Korea', 'heat-China', 
                   
                   'humidity-France', 'humidity-Germany', 'humidity-India', 'humidity-Indonesia', 'humidity-Italy', 'humidity-Japan', 'humidity-Korea', 
                   'humidity-China', 
                   
                   'temperature-France', 'temperature-Germany', 'temperature-India', 'temperature-Indonesia', 'temperature-Italy', 'temperature-Japan', 
                   'temperature-Korea', 'temperature-China', 
                   
                   'precipitation-France', 'precipitation-Germany', 'precipitation-India', 'precipitation-Indonesia', 'precipitation-Italy', 'precipitation-Japan', 
                   'precipitation-Korea', 'precipitation-China']

# 5. FamaFrench 5 factors and climate 5 factors with XGBRegressor

## example with AAPL

X = all_data[features]
y = all_data['ret_AAPL'] - all_data['RF']    # you can change only ticker from stock list
#stock_list = ['AAPL',  'APTV', 'BAC', 'BALL',  'BEP', 'BKNG', 'BMRN', 'COST', 'CVS', 'DE', 'ETN', 'GOOGL', 'HD', 'HIG',  'MS', 'MSFT', 'PLD',  'TT', 'UNH', 'V']

# X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.25, random_state=42)

%%time
params={'gamma':[0, 0.01, 0.05, 0.1, 0.5, 1, 2, 3], 
        'colsample_bylevel':[0.9], 
        'colsample_bytree':[0.8], 
        'colsample_bynode':[0.5], 
        'max_depth':np.arange(4, 10 ,2), 
        'n_estimators':[100,200],
        'subsample':[0.5, 0.6, 0.7, 0.8, 0.9, 1], 
        'min_child_weight':[1, 2, 3, 4, 5], 
        'learning_rate':[0.1, 0.2, 0.3, 0.4, 0.5]}

xgb_reg = GridSearchCV(XGBRegressor(booster = 'gbtree'), params, cv=5, n_jobs=-1)

grid_result = xgb_reg.fit(X, y)
print('Best Score: ', grid_result.best_score_)
print('Best Params: ', grid_result.best_params_)

y_pred = xgb_reg.predict(X)

print('mean_squared_error =\t\t', mean_squared_error(y, y_pred))
print('mean_absolute_error =\t\t', mean_absolute_error(y, y_pred))
