==> 6. ESG_trend.ipynb <==
from pytrends.request import TrendReq
#from pytrendsasync.request import TrendReq
import pandas as pd

Download ESG Glossary from here : https://breckinridge-fs.s3.amazonaws.com/files/uploads/esg-glossary-2.pdf

glossary_list = ['B Corp',
                'Benefit Corporation',
                'Best in Class',
                'Biodiversity',
                'Biofuel',
                'Carbon Footprint',
                'Carbon Neutral',
                'Carbon Pricing',
                'CDP',
                'Ceres',
                'Chief Executives for Corporate ',
                'Purpose (CECP)',
                'Climate Action 100+',
                'Climate Change',
                'Clean Energy/Renewable Energy',
                'Circular Economy',
                'Conference of the Parties (COP)',
                'Corporate Social Responsibility',
                'CSR Report',
                'Diversity, Equity, and Inclusion',
                'Divestment',
                'Environmental Factors',
                'Environmental, Social and ',
                'Governance (ESG)',
                'ESG Integration',
                'Ethical Investing',
                'EU Taxonomy (for Sustainable ',
                'Activities)',
                'Exclusionary (or Negative) Screening',
                'Extra-financial Risks',
                'Gender Equity',
                'Gender Lens Investing',
                'Global Reporting Initiative (GRI)',
                'Governance Factors',
                'Green Bonds',
                'Greenhouse Gas (GHG) Emissions',
                'GHG Protocol',
                'Greenwashing',
                'Impact Investing',
                'Integrated Reporting',
                'Issuer Engagement',
                'Materiality',
                'Net Zero',
                'Physical Climate Risks',
                'Positive Screening',
                'Science-Based Targets Initiative (SBTi)',
                'Scope 1 - Under the GHG Protocol, ',
                'All Direct Emissions',
                'Scope 2 - Under the GHG Protocol, ',
                'Indirect Emissions',
                'Scope 3 - Under the GHG Protocol, ',
                'All Other Indirect Emissions',
                'Shareholder Advocacy',
                'Sin Stocks',
                'Social Bonds',
                'Social Factors',
                'Social Progress Index (SPI)',
                'Socially Responsible Investing (SRI)',
                'Stakeholder Engagement',
                'Stranded Assets',
                'Sustainability Bonds',
                'Sustainability-Linked Bonds',
                'Sustainable Investing',
                'The Sustainability Accounting ',
                'Standards Board (SASB)',
                'Sustainable Development Goals (SDG)',
                'Sustainable Finance Disclosure ',
                'Regulation (SFDR)',
                'Sustainable Investing',
                'Task Force on Climate-Related ',
                'Financial Disclosures (TCFD)',
                'Thematic Investing',
                'Transition Risk',
                'Triple Bottom Line',
                'United Nations Framework Convention ',
                'on Climate Change (UNFCCC)',
                'United Nations Principles for ',
                'Responsible Investing (UN PRI)',
                'Values-Based and Impact Investing',
                'Zero Waste']

len(glossary_list)

pytrend = TrendReq()
df = pd.DataFrame()
for i in range(1,81):
    pytrend.build_payload(kw_list=glossary_list[i-1 : i] , timeframe='2015-09-30 2023-03-03')
    dg = pytrend.interest_over_time()
    #dg = dg.drop(labels=['isPartial'],axis='columns')
    #dg = pd.DataFrame(dg)
    print (dg)
    #print (dg.head(10))
    df = pd.concat([df, dg], axis=1)
    #df = pd.concat([df, dg.values()], axis=1)
    #df = pd.merge(df, dg)
df = df.drop(columns=[ 'isPartial'])
df = df.loc[(df.sum(axis=1) != 0), (df.sum(axis=0) != 0)]

df.head()

df.info()

# please save df to csv
