==> 7. Climate_Score_each_Stock.ipynb <==
# Initial Imports:
import pandas as pd
import numpy as np
import os
from datetime import datetime

# To run models:
from xgboost import XGBRegressor, XGBClassifier
from sklearn.metrics import (r2_score, 
                             ConfusionMatrixDisplay, 
                             accuracy_score, 
                             precision_score, 
                             recall_score, 
                             f1_score)

BASE_DIR = os.path.abspath(os.path.dirname('__file__'))

all_data  = pd.read_csv('all_data.csv', index_col='Date', parse_dates=True)

all_data.head()

all_data = all_data.iloc[1:,]
all_data = all_data.ffill()

stock_list = ['ret_AAPL', 'ret_APTV', 'ret_BAC', 'ret_BALL', 'ret_BEP', 'ret_BKNG', 
              'ret_BMRN', 'ret_COST', 'ret_CVS', 'ret_DE', 'ret_ETN', 'ret_GOOGL', 'ret_HD', 
              'ret_HIG', 'ret_MS', 'ret_MSFT', 'ret_PLD', 'ret_TT', 'ret_UNH', 'ret_V']

for i in stock_list:
    all_data[i] = np.where(all_data[i] > 0, 1, 0) # 'Long = 1', 'Short = -1'

# example : Create C_Score for AAPL

### Note! C_score is climate score for each stock

# Define X and y variables:
X = all_data.drop(stock_list, axis=1)
X = X.drop('ret_Fund' ,axis=1)
y = all_data.loc[:, 'ret_AAPL']

# Split into Training/Testing Data:
split = int(0.8 * len(X))
X_train = X[: split]
X_test = X[split:]
y_train = y[: split]
y_test = y[split:]

clf = XGBClassifier(booster = 'gbtree', 
                    objective = 'binary:logistic',
                    gamma = 0.05,
                    colsample_bytree = 0.8,
                    colsample_bynode = 0.5,
                    n_estimators = 200, 
                    subsample = 0.8, 
                    min_child_weight = 1, 
                    learning_rate = 0.1)

%%time
clf.fit(X_train, y_train)

predictions = clf.predict(X)

accuracy_score(y, predictions)

precision_score(y, predictions, pos_label=0)

precision_score(y, predictions, pos_label=1)

recall_score(y, predictions, pos_label=0)

recall_score(y, predictions, pos_label=1)

# Calculate C_Score for all stock

stock_list = ['AAPL', 'APTV', 'BAC', 'BALL', 'BEP', 'BKNG', 'BMRN', 'COST', 'CVS', 'DE', 'ETN', 'GOOGL', 'HD', 'HIG', 'MS', 'MSFT', 'PLD', 'TT', 'UNH', 'V']

%%time
for i in stock_list:
    # Define X and y variables:
    X = all_data.drop(stock_list, axis=1)
    X = X.drop('ret_Fund' ,axis=1)
    y = all_data.loc[:, 'ret_'+i]

    # Split into Training/Testing Data:
    split = int(0.8 * len(X))
    X_train = X[: split]
    X_test = X[split:]
    y_train = y[: split]
    y_test = y[split:]

    clf = XGBClassifier(booster = 'gbtree', 
                        objective = 'binary:logistic',
                        gamma = 0.05,
                        colsample_bytree = 0.8,
                        colsample_bynode = 0.5,
                        n_estimators = 200, 
                        subsample = 0.8, 
                        min_child_weight = 1, 
                        learning_rate = 0.1)
    clf.fit(X_train, y_train)
    predictions = clf.predict(X)
    all_data['C_Score_'+i] = predictions

all_data.head()

score  = all_data.columns[[x.startswith("C_Score") for x in all_data.columns]]
c_score = all_data[score]

c_score
